ai-service/requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt

python-consul
requests
ultralytics
opencv-python-headless
mysqlclient